# geox-api/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Mercator API",
    "contact": {
      "email": "duber+ManagementApi@mercator.tech"
    },
    "version": "0.0.1"
  },
  "paths": {
...
    "/geofencer/shapes/export": {
      "post": {
        "tags": [
          "geofencer"
        ],
        "summary": "Shapes Export",
        "description": "Export shapes to S3.\n\nThis is an async task. Use `/tasks/results/{task_id}` to retrieve the status and results.",
        "operationId": "shapes_export_geofencer_shapes_export_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CeleryTaskResponse"
                }
              }
            }
          },
          "501": {
            "description": "Data export not supported on the server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/shapes/export": {
      "post": {
        "tags": [
          "geofencer"
        ],
        "summary": "Shapes Export",
        "description": "Export shapes to S3.\n\nThis is an async task. Use `/tasks/results/{task_id}` to retrieve the status and results.",
        "operationId": "shapes_export_shapes_export_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CeleryTaskResponse"
                }
              }
            }
          },
          "501": {
            "description": "Data export not supported on the server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorModel"
                }
              }
            }
          }
        },
        "deprecated": true,
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/geofencer/shapes/bulk": {
      "post": {
        "tags": [
          "geofencer"
        ],
        "summary": "Bulk Create Shapes",
        "description": "Create multiple shapes.",
        "operationId": "bulk_create_shapes_geofencer_shapes_bulk_post",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Namespace",
              "type": "string",
              "format": "uuid4"
            },
            "name": "namespace",
            "in": "query"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "Data",
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/GeoShapeCreate"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShapeCountResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "geofencer"
        ],
        "summary": "Bulk Delete Shapes",
        "description": "Create multiple shapes.",
        "operationId": "bulk_delete_shapes_geofencer_shapes_bulk_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "Shape Uuids",
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid4"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShapeCountResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/geofencer/shapes": {
      "get": {
        "tags": [
          "geofencer"
        ],
        "summary": " Get Shapes",
        "description": "Read shapes.\n\nWill return 200 even if no shapes match the query, including the case in which the\nnamespace does not exist.",
        "operationId": "_get_shapes_geofencer_shapes_get",
        "parameters": [
          {
            "description": "Only include shapes in the specified namespace given its UUID or name.",
            "required": false,
            "schema": {
              "title": "Namespace of shapes",
              "type": "string",
              "description": "Only include shapes in the specified namespace given its UUID or name.",
              "format": "uuid4"
            },
            "name": "namespace",
            "in": "query"
          },
          {
            "description": "If TRUE, then only return shapes of the requesting user.",
            "required": false,
            "schema": {
              "title": "Get shapes for a user",
              "type": "boolean",
              "description": "If TRUE, then only return shapes of the requesting user."
            },
            "name": "user",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Item offset",
              "minimum": 0.0,
              "type": "integer",
              "default": 0
            },
            "name": "offset",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Number of shapes to retrieve",
              "minimum": 1.0,
              "type": "integer",
              "default": 300
            },
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response  Get Shapes Geofencer Shapes Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GeoShape"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "geofencer"
        ],
        "summary": " Post Shapes",
        "description": "Create a shape.",
        "operationId": "_post_shapes_geofencer_shapes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeoShapeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoShape"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/geofencer/shapes/op/count": {
      "get": {
        "tags": [
          "geofencer"
        ],
        "summary": "Get Shape Count",
        "description": "Get shape count.",
        "operationId": "get_shape_count_geofencer_shapes_op_count_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShapeCountResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/geofencer/shapes/op/contains": {
      "get": {
        "tags": [
          "geofencer"
        ],
        "summary": "Get Shapes Containing Point",
        "description": "Get shapes containing a point.",
        "operationId": "get_shapes_containing_point_geofencer_shapes_op_contains_get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Lat",
              "maximum": 90.0,
              "minimum": -90.0,
              "type": "number"
            },
            "name": "lat",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "Lng",
              "maximum": 180.0,
              "minimum": -180.0,
              "type": "number"
            },
            "name": "lng",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Shapes Containing Point Geofencer Shapes Op Contains Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Feature"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/geofencer/shapes/{shape_id}": {
      "get": {
        "tags": [
          "geofencer"
        ],
        "summary": " Get Shapes  Shape Id",
        "description": "Read a shape.",
        "operationId": "_get_shapes__shape_id_geofencer_shapes__shape_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Shape Id",
              "type": "string",
              "format": "uuid4"
            },
            "name": "shape_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoShape"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "geofencer"
        ],
        "summary": " Delete Shapes  Shape Id",
        "description": "Delete a shape.",
        "operationId": "_delete_shapes__shape_id_geofencer_shapes__shape_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Shape Id",
              "type": "string",
              "format": "uuid4"
            },
            "name": "shape_id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Shape does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "geofencer"
        ],
        "summary": " Patch Shapes  Shape Id",
        "description": "Update a shape.",
        "operationId": "_patch_shapes__shape_id_geofencer_shapes__shape_id__patch",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "ID of the shape to edit.",
              "type": "string",
              "format": "uuid4"
            },
            "name": "shape_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeoShapeUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoShape"
                }
              }
            }
          },
          "404": {
            "description": "Shape does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/geofencer/shapes/{uuid}": {
      "put": {
        "tags": [
          "geofencer"
        ],
        "summary": " Update Shapes  Shape Id",
        "description": "Update a shape.",
        "operationId": "_update_shapes__shape_id_geofencer_shapes__uuid__put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeoShapeUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoShape"
                }
              }
            }
          },
          "404": {
            "description": "Shape does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestErrorModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "deprecated": true,
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/geofencer/shapes/op/{operation}": {
      "post": {
        "tags": [
          "geofencer"
        ],
        "summary": "Get Shapes By Operation",
        "description": "Get shapes by operation.",
        "operationId": "get_shapes_by_operation_geofencer_shapes_op__operation__post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/GeometryOperation"
            },
            "name": "operation",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "Geom",
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/Point"
                  },
                  {
                    "$ref": "#/components/schemas/Polygon"
                  },
                  {
                    "$ref": "#/components/schemas/LineString"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Shapes By Operation Geofencer Shapes Op  Operation  Post",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Feature"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Operation not enabled for this account"
          },
          "501": {
            "description": "Operation not supported on the server."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/tasks/results/{task_id}": {
      "get": {
        "tags": [
          "geofencer",
          "tasks"
        ],
        "summary": "Get Status",
        "description": "Retrieve results of a task.",
        "operationId": "get_status_tasks_results__task_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            },
            "name": "task_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CeleryTaskResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CeleryTaskResponse": {
        "title": "CeleryTaskResponse",
        "required": [
          "task_id"
        ],
        "type": "object",
        "properties": {
          "task_id": {
            "title": "Task Id",
            "type": "string",
            "description": "Task id."
          }
        },
        "description": "Response from submitting a Celery task."
      },
...
      "Feature": {
        "title": "Feature",
        "type": "object",
        "properties": {
          "type": {
            "title": "Type",
            "type": "string",
            "default": "Feature",
            "const": "Feature"
          },
          "geometry": {
            "title": "Geometry",
            "anyOf": [
              {
                "$ref": "#/components/schemas/Point"
              },
              {
                "$ref": "#/components/schemas/MultiPoint"
              },
              {
                "$ref": "#/components/schemas/LineString"
              },
              {
                "$ref": "#/components/schemas/MultiLineString"
              },
              {
                "$ref": "#/components/schemas/Polygon"
              },
              {
                "$ref": "#/components/schemas/MultiPolygon"
              },
              {
                "$ref": "#/components/schemas/GeometryCollection"
              }
            ]
          },
          "properties": {
            "title": "Properties",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "bbox": {
            "title": "Bbox",
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "description": "Feature Model"
      },
...
      "GeoShapeCreate": {
        "title": "GeoShapeCreate",
        "required": [
          "geojson"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string",
            "description": "Name of the shape"
          },
          "geojson": {
            "title": "Geojson",
            "allOf": [
              {
                "$ref": "#/components/schemas/Feature"
              }
            ],
            "description": "GeoJSON representation of the shape"
          },
          "namespace": {
            "title": "Namespace",
            "type": "string",
            "description": "Namespace id.",
            "format": "uuid4"
          }
        },
        "example": {
          "name": "Lake Merritt",
          "properties": {},
          "namespace_id": "cf9f83cd-d454-4abb-b77c-c6d639804618",
          "geojson": {
            "type": "Feature",
            "properties": {},
            "geometry": {
              "type": "Polygon",
              "coordinates": [
                [
                  [
                    -122.26907730102539,
                    37.796084948503236
                  ],
                  [
                    -122.24555969238281,
                    37.796084948503236
                  ],
                  [
                    -122.24555969238281,
                    37.81208947621168
                  ],
                  [
                    -122.26907730102539,
                    37.81208947621168
                  ],
                  [
                    -122.26907730102539,
                    37.796084948503236
                  ]
                ]
              ]
            }
          }
        }
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "RequestErrorModel": {
        "title": "RequestErrorModel",
        "required": [
          "detail"
        ],
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        }
      },
      "ShapeCountResponse": {
        "title": "ShapeCountResponse",
        "required": [
          "num_shapes"
        ],
        "type": "object",
        "properties": {
          "num_shapes": {
            "title": "Num Shapes",
            "type": "integer",
            "description": "Number of shapes affected by transaction"
          }
        },
        "example": {
          "num_shapes": 5
        }
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}